backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.4
psycopg2-binary>=2.9
Pillow>=12.0
django-environ>=0.11

drf-spectacular>=0.27


